# Conflicts:
#	QuanLyDiemSinhVien/BUL/BUL.csproj
#	QuanLyDiemSinhVien/Common/Constant.cs
#	QuanLyDiemSinhVien/DAL/DataProvider.cs
#	QuanLyDiemSinhVien/DAL/GiangVienDAL.cs
#	QuanLyDiemSinhVien/DTO/DTO.csproj
#	QuanLyDiemSinhVien/QuanLyDiemSinhVien/frmMain.Designer.cs
#	QuanLyDiemSinhVien/QuanLyDiemSinhVien/frmMain.cs
#	QuanLyDiemSinhVien/QuanLyDiemSinhVien/frmMain.resx

+++ QuanLyDiemSinhVien/QuanLyDiemSinhVien/MonHocGUI/frmMonHoc.cs
@@ -17,6 +17,7 @@ namespace QuanLyDiemSinhVien.MonHocGUI
     {
         #region Fields
         private List<MonHocDTO> m_lstMonHoc;
+        private BindingSource m_bdsMonHoc;
         #endregion
 
         #region Constructor
@@ -29,31 +30,19 @@ namespace QuanLyDiemSinhVien.MonHocGUI
         #region UI Event
         private void frmMonHoc_Load(object sender, EventArgs e)
         {
-            m_lstMonHoc = MonHocBUL.LoadMonHoc();
-            if (m_lstMonHoc == null)
-            {
-                return;
-            }
-
-            dgvMonHoc.DataSource = m_lstMonHoc;
-            InitializeDataGridView();
+            m_bdsMonHoc = new BindingSource();
+            ReloadGUI();
         }
 
         private void dgvMonHoc_Click(object sender, EventArgs e)
         {
-            if (dgvMonHoc.SelectedRows.Count == 0)
-            {
-                return;
-            }
-            DataGridViewRow dgvRow = dgvMonHoc.SelectedRows[0];
-            txtMaMonHoc.Text = dgvRow.Cells["MaMH"].Value.ToString();
-            txtTenMonHoc.Text = dgvRow.Cells["TenMH"].Value.ToString();
+            SetupTextBox();
         }
 
         private void btnThemMonHoc_Click(object sender, EventArgs e)
         {
-            string strMaMonHoc = txtMaMonHoc.Text;
-            string strTenMonHoc = txtTenMonHoc.Text;
+            string strMaMonHoc = txtMaMonHoc.Text.Trim();
+            string strTenMonHoc = txtTenMonHoc.Text.Trim();
 
             try
             {
@@ -81,14 +70,14 @@ namespace QuanLyDiemSinhVien.MonHocGUI
             {
                 MessageBox.Show("Thêm môn học thành công", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 AddMonHocIntoList(strMaMonHoc, strTenMonHoc);
-                ReloadDataGridView();
+                ReloadGUI();
             }
         }
 
         private void btnCapNhatMonHoc_Click(object sender, EventArgs e)
         {
-            string strMaMonHoc = txtMaMonHoc.Text;
-            string strTenMonHoc = txtTenMonHoc.Text;
+            string strMaMonHoc = txtMaMonHoc.Text.Trim();
+            string strTenMonHoc = txtTenMonHoc.Text.Trim();
 
             try
             {
@@ -100,14 +89,18 @@ namespace QuanLyDiemSinhVien.MonHocGUI
                 return;
             }
 
-            if (MonHocBUL.UpdateMonHoc(txtMaMonHoc.Text, txtTenMonHoc.Text))
+            if (MonHocBUL.UpdateMonHoc(strMaMonHoc, strTenMonHoc))
             {
-                MessageBox.Show("Cập nhật thành công", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                MessageBox.Show("Cập nhật môn học thành công", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                UpdateMonHocIntoList(strMaMonHoc, strTenMonHoc);
+                ReloadGUI();
             }
         }
 
         private void btnXoaBoMonHoc_Click(object sender, EventArgs e)
         {
+            string strMaMonHoc = txtMaMonHoc.Text.Trim();
+
             try
             {
                 ValidateInput();
@@ -123,10 +116,23 @@ namespace QuanLyDiemSinhVien.MonHocGUI
                 return;
             }
 
+            if (MonHocBUL.IsExistedAtTableDiem(strMaMonHoc))
+            {
+                MessageBox.Show("Môn học này không được xóa", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+
             if (MonHocBUL.DeleteMonHoc(txtMaMonHoc.Text.Trim()))
             {
                 MessageBox.Show("Xóa môn học thành công", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                DeleteMonHocIntoList(strMaMonHoc);
+                ReloadGUI();
             }
+        }
+
+        private void btnTaiLai_Click(object sender, EventArgs e)
+        {
+            ReloadGUI();
         }
 
         private void btnPhucHoiMonHoc_Click(object sender, EventArgs e)
@@ -166,12 +172,6 @@ namespace QuanLyDiemSinhVien.MonHocGUI
             return false;
         }
 
-        private void ReloadDataGridView()
-        {
-            dgvMonHoc.DataSource = typeof(List<MonHocDTO>);
-            dgvMonHoc.DataSource = m_lstMonHoc;
-        }
-
         private void AddMonHocIntoList(string x_strMaMonHoc, string x_strTenMonHoc)
         {
             MonHocDTO objMonHoc = new MonHocDTO();
@@ -182,26 +182,105 @@ namespace QuanLyDiemSinhVien.MonHocGUI
 
         private void UpdateMonHocIntoList(string x_strMaMonHoc, string x_strTenMonHoc)
         {
-            MonHocDTO objMonHoc = new MonHocDTO();
-            objMonHoc.MaMH = x_strMaMonHoc;
-            objMonHoc.TenMH = x_strTenMonHoc;
-            m_lstMonHoc.Add(objMonHoc);
+            for(int i = 0; i<m_lstMonHoc.Count; i++)
+            {
+                if (m_lstMonHoc[i].MaMH.Equals(x_strMaMonHoc.Trim()))
+                {
+                    m_lstMonHoc[i].TenMH = x_strTenMonHoc.Trim();
+                    return;
+                }
+            }
+        }
+
+        private void DeleteMonHocIntoList(string x_strMaMonHoc)
+        {
+            for (int i = 0; i < m_lstMonHoc.Count; i++)
+            {
+                if (m_lstMonHoc[i].MaMH.Equals(x_strMaMonHoc.Trim()))
+                {
+                    try
+                    {
+                        m_lstMonHoc.RemoveAt(i);
+                    }
+                    catch
+                    {
+
+                    }
+                }
+            }
         }
 
         private void ValidateInput()
         {
-            string strMaMonHoc = txtMaMonHoc.Text;
-            string strTenMonHoc = txtTenMonHoc.Text;
+            string strMaMonHoc = txtMaMonHoc.Text.Trim();
+            string strTenMonHoc = txtTenMonHoc.Text.Trim();
 
             if (string.IsNullOrWhiteSpace(strMaMonHoc))
             {
-                throw new Exception("MÃ MÔN HỌC không được để trống");
+                throw new Exception("MÃ MÔN HỌC không được để trống!");
             }
 
             if (string.IsNullOrWhiteSpace(strTenMonHoc))
             {
-                throw new Exception("TÊN MÔN HỌC không được để trống");
+                throw new Exception("TÊN MÔN HỌC không được để trống!");
             }
+
+            if(strMaMonHoc.Length > 8)
+            {
+                throw new Exception("MÃ MÔN HỌC quá dài!");
+            }
+
+            if (strTenMonHoc.Length > 40)
+            {
+                throw new Exception("MÃ MÔN HỌC quá dài!");
+            }
+        }
+
+        private void ReloadGUI()
+        {
+            LoadDataGridView();
+            InitializeDataGridView();
+            SetupTextBox();
+        }
+
+        private void LoadDataGridView()
+        {
+            m_lstMonHoc = MonHocBUL.LoadMonHoc();
+            if (m_lstMonHoc == null)
+            {
+                return;
+            }
+            m_bdsMonHoc.DataSource = m_lstMonHoc;
+            m_bdsMonHoc.Sort = "MAMH ASC";
+            dgvMonHoc.DataSource = typeof(BindingSource);
+            dgvMonHoc.DataSource = m_bdsMonHoc;
+        }
+
+        private void SetupTextBox()
+        {
+            DataGridViewRow dgvRow = dgvMonHoc.SelectedRows[0];
+            if (dgvMonHoc.Rows.Count - 1 == dgvRow.Index)
+            {
+                txtMaMonHoc.Text = string.Empty;
+                txtTenMonHoc.Text = string.Empty;
+                txtMaMonHoc.Enabled = true;
+                txtTenMonHoc.Enabled = true;
+
+                btnThemMonHoc.Enabled = true;
+                btnCapNhatMonHoc.Enabled = false;
+                btnXoaBoMonHoc.Enabled = false;
+
+                return;
+            }
+
+            txtMaMonHoc.Text = dgvRow.Cells[0].Value.ToString().Trim();
+            txtTenMonHoc.Text = dgvRow.Cells[1].Value.ToString().Trim();
+            txtMaMonHoc.Enabled = false;
+            txtTenMonHoc.Enabled = true;
+
+            btnThemMonHoc.Enabled = false;
+            btnCapNhatMonHoc.Enabled = true;
+            btnXoaBoMonHoc.Enabled = true;
         }
         #endregion
     }
